refactores ItemCard to have a changeable header
