+++ components/ItemCard/index.js
@@ -8,31 +8,12 @@ import CardImage from "./CardImage.js";
 import CardDescription from "./CardDescription.js";
 import UsernameButton from "./UsernameButton.js";
 
-const ItemCard = ({ id }) => {
-  const [isFavorited, setFavorited] = useState(false);
-
-  const handleFavoritedItem = () => {
-    setFavorited(() => !isFavorited);
-  };
+const ItemCard = ({ children, id }) => {
 
   return (
     <Link href={`/products/${id}`}>
       <Card raised sx={{ width: "15vw", cursor: "pointer" }}>
-        <CardHeader
-          sx={{ zIndex: 9999 }}
-          action={
-            <IconButton
-              aria-label="add to favorites"
-              onClick={handleFavoritedItem}
-            >
-              {!isFavorited ? (
-                <FavoriteBorderIcon />
-              ) : (
-                <FavoriteIcon sx={{ color: "red" }} />
-              )}
-            </IconButton>
-          }
-        />
+        {children}
         <CardImage />
         <CardDescription />
         <UsernameButton />
